fix: rename `util.h` to avoid build error
`util.h` is included by firmware files causing a build error.

The error only manifests when rebuilding the firmware (e.g. new build or config change).
