+++ layouts/main/keymap.c
@@ -14,7 +14,7 @@
 	#define output(...) (void)0
 #endif
 
-#include "util.h"
+#include "utilities.h"
 
 #include "leds.h"
 #include "overlay.h"
